moved initial capacitor stuff to the app.vue file
