+++ pages/login.vue
@@ -4,6 +4,7 @@ import VLoginWithProvider from '@nypublicradio/nypr-design-system-vue3/v2/src/co
 import VFlexibleLink from '@nypublicradio/nypr-design-system-vue3/v2/src/components/VFlexibleLink.vue'
 
 definePageMeta({
+  layout: 'blank',
   pageTransition: {
     name: 'login',
   },
